# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crack402 Sandbox API",
    "version": "1.0.0",
    "description": "API for managing cloud sandboxes with x402 payment support"
  },
  "servers": [
    {
      "url": "https://api.crack402.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "TenantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Tenant API key for sandbox operations"
      },
      "NodeToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Node bootstrap token (Bearer)"
      },
      "OperatorApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Operator-Key",
        "description": "Operator API key for management operations"
      }
    },
    "schemas": {
      "Sandbox": {
        "type": "object",
        "properties": {
          "sandbox_id": {
            "type": "string",
            "description": "Unique sandbox identifier",
            "example": "sbx_abc12345678901234567"
          },
          "status": {
            "$ref": "#/components/schemas/SandboxStatus"
          },
          "base_image": {
            "type": "string",
            "description": "Base container image",
            "example": "ubuntu:22.04"
          },
          "vcpu": {
            "type": "number",
            "description": "Allocated virtual CPUs",
            "example": 2
          },
          "memory_mb": {
            "type": "integer",
            "description": "Allocated memory in MB",
            "example": 2048
          },
          "gpu": {
            "type": "string",
            "nullable": true,
            "description": "GPU type or null"
          },
          "timeout_seconds": {
            "type": "integer",
            "description": "Max sandbox lifetime in seconds",
            "example": 3600
          },
          "idle_timeout_seconds": {
            "type": "integer",
            "description": "Idle timeout in seconds",
            "example": 600
          },
          "network_policy": {
            "type": "string",
            "description": "Network access policy",
            "example": "outbound-only"
          },
          "env_vars": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Environment variables"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Arbitrary key-value metadata",
            "example": {
              "env": "production",
              "team": "backend"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 creation timestamp",
            "example": "2026-01-15T12:00:00Z"
          },
          "started_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "ISO 8601 start timestamp, null if not started"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 expiration timestamp"
          },
          "public_url": {
            "type": "string",
            "nullable": true,
            "description": "Public URL for the sandbox",
            "example": "https://sbx-abc123.sandbox.example.com"
          },
          "private_ip": {
            "type": "string",
            "nullable": true,
            "description": "Private IP address",
            "example": "10.0.1.5"
          },
          "region": {
            "type": "string",
            "description": "Deployment region",
            "example": "us-east-1"
          },
          "isolation_backend": {
            "$ref": "#/components/schemas/IsolationBackend"
          },
          "cost_accrued_usd": {
            "type": "number",
            "description": "Total cost accrued in USD",
            "example": 0.42
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SandboxEvent"
            },
            "description": "Recent sandbox events"
          }
        },
        "required": [
          "sandbox_id",
          "status",
          "base_image",
          "vcpu",
          "memory_mb",
          "gpu",
          "timeout_seconds",
          "idle_timeout_seconds",
          "network_policy",
          "created_at",
          "started_at"
        ]
      },
      "SandboxStatus": {
        "type": "string",
        "enum": [
          "provisioning",
          "ready",
          "running",
          "paused",
          "stopping",
          "stopped",
          "error",
          "destroyed"
        ],
        "description": "Current sandbox status"
      },
      "IsolationBackend": {
        "type": "string",
        "enum": [
          "firecracker",
          "gvisor",
          "kata",
          "wasm",
          "container",
          "v8-isolate"
        ],
        "description": "Isolation technology used"
      },
      "SandboxEvent": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Event type",
            "example": "state_change"
          },
          "status": {
            "type": "string",
            "description": "Sandbox status at event time",
            "example": "running"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 event timestamp",
            "example": "2026-01-15T12:00:00Z"
          }
        },
        "required": [
          "type",
          "status",
          "timestamp"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "The requested sandbox does not exist"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Additional error context"
          },
          "request_id": {
            "type": "string",
            "description": "Request ID for support reference",
            "example": "req_abc123"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "invalid_request",
          "unauthorized",
          "forbidden",
          "sandbox_not_found",
          "sandbox_state_conflict",
          "exec_timeout",
          "resource_limit_exceeded",
          "capacity_exhausted",
          "payment_required",
          "rate_limited",
          "internal_error"
        ],
        "description": "Machine-readable error code",
        "example": "sandbox_not_found"
      },
      "CreateSandboxRequest": {
        "type": "object",
        "properties": {
          "base_image": {
            "type": "string",
            "description": "Base container image",
            "example": "ubuntu:22.04"
          },
          "vcpu": {
            "type": "number",
            "minimum": 0.25,
            "maximum": 32,
            "default": 2,
            "description": "Number of virtual CPUs",
            "example": 2
          },
          "memory_mb": {
            "type": "integer",
            "minimum": 256,
            "maximum": 131072,
            "default": 2048,
            "description": "Memory allocation in MB",
            "example": 2048
          },
          "gpu": {
            "type": "string",
            "nullable": true,
            "enum": [
              "T4",
              "L4",
              "A10G",
              "L40S",
              "A100-40",
              "A100-80",
              "H100",
              "H200",
              "B200",
              null
            ],
            "description": "GPU type, null for no GPU"
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 10,
            "maximum": 86400,
            "default": 3600,
            "description": "Max sandbox lifetime in seconds",
            "example": 3600
          },
          "idle_timeout_seconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 86400,
            "default": 600,
            "description": "Idle timeout before auto-pause/destroy in seconds",
            "example": 600
          },
          "auto_destroy": {
            "type": "boolean",
            "default": true,
            "description": "Destroy sandbox when timeout expires",
            "example": true
          },
          "auto_pause_on_idle": {
            "type": "boolean",
            "default": false,
            "description": "Pause sandbox when idle instead of destroying",
            "example": false
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "Inline code to execute on start"
          },
          "language": {
            "type": "string",
            "nullable": true,
            "enum": [
              "python",
              "javascript",
              "typescript",
              "ruby",
              "go",
              "rust",
              "java",
              "bash",
              null
            ],
            "description": "Language for inline code execution"
          },
          "github_repo": {
            "type": "object",
            "nullable": true,
            "properties": {
              "url": {
                "type": "string",
                "format": "uri",
                "description": "GitHub repository URL",
                "example": "https://github.com/org/repo"
              },
              "branch": {
                "type": "string",
                "default": "main",
                "description": "Branch to clone",
                "example": "main"
              },
              "commit": {
                "type": "string",
                "description": "Specific commit SHA to checkout",
                "example": "a1b2c3d4"
              },
              "path": {
                "type": "string",
                "default": "/workspace",
                "description": "Path to clone into",
                "example": "/workspace"
              },
              "token": {
                "type": "string",
                "description": "GitHub access token for private repos"
              }
            },
            "required": [
              "url"
            ],
            "description": "GitHub repo to clone into the sandbox"
          },
          "env_vars": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Environment variables to set",
            "example": {
              "NODE_ENV": "production"
            }
          },
          "network_policy": {
            "$ref": "#/components/schemas/NetworkPolicy"
          },
          "exposed_ports": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Ports to expose publicly",
            "example": [
              3000,
              8080
            ]
          },
          "volumes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Volume name",
                  "example": "data"
                },
                "mount_path": {
                  "type": "string",
                  "description": "Mount path inside sandbox",
                  "example": "/data"
                },
                "size_gb": {
                  "type": "number",
                  "description": "Volume size in GB",
                  "example": 10
                }
              },
              "required": [
                "name",
                "mount_path",
                "size_gb"
              ]
            },
            "description": "Persistent volumes to attach"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Arbitrary key-value metadata",
            "example": {
              "env": "production",
              "team": "backend"
            }
          },
          "payment": {
            "type": "object",
            "nullable": true,
            "properties": {
              "method": {
                "type": "string",
                "enum": [
                  "x402",
                  "api_key_billing",
                  "prepaid_credits"
                ],
                "description": "Payment method",
                "example": "api_key_billing"
              },
              "x402_signature": {
                "type": "string",
                "description": "x402 payment signature"
              },
              "budget_limit_usd": {
                "type": "number",
                "description": "Maximum budget in USD",
                "example": 10
              }
            },
            "required": [
              "method"
            ],
            "description": "Payment information"
          }
        },
        "required": [
          "base_image"
        ]
      },
      "NetworkPolicy": {
        "type": "string",
        "enum": [
          "none",
          "outbound-only",
          "full"
        ],
        "default": "outbound-only",
        "description": "Network access policy",
        "example": "outbound-only"
      },
...
      "WriteFileRequest": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "Absolute file path inside the sandbox",
            "example": "/workspace/main.py"
          },
          "content": {
            "type": "string",
            "description": "File content"
          },
          "encoding": {
            "type": "string",
            "enum": [
              "utf-8",
              "base64"
            ],
            "default": "utf-8",
            "description": "Content encoding",
            "example": "utf-8"
          },
          "permissions": {
            "type": "string",
            "description": "Unix file permissions",
            "example": "0644"
          }
        },
        "required": [
          "path",
          "content"
        ]
      },
      "FileListResponse": {
        "type": "object",
        "properties": {
          "directory": {
            "type": "string",
            "description": "Listed directory path",
            "example": "/workspace"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "description": "Files and directories in the listed path"
          }
        },
        "required": [
          "directory",
          "files"
        ]
      },
      "FileEntry": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "File or directory path",
            "example": "/workspace/src/index.ts"
          },
          "type": {
            "type": "string",
            "enum": [
              "file",
              "directory",
              "symlink"
            ],
            "description": "Entry type",
            "example": "file"
          },
          "size_bytes": {
            "type": "integer",
            "description": "Size in bytes (0 for directories)",
            "example": 2048
          },
          "modified_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 last modified timestamp",
            "example": "2026-01-15T12:00:00Z"
          }
        },
        "required": [
          "path",
          "type",
          "size_bytes",
          "modified_at"
        ]
      },
...
    },
...
  },
  "paths": {
    "/v1/sandboxes": {
      "post": {
        "tags": [
          "Sandboxes"
        ],
        "summary": "Create a new sandbox",
        "description": "Provision a new sandbox with the specified configuration.",
        "security": [
          {
            "TenantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sandbox created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Resource limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No capacity available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sandboxes"
        ],
        "summary": "List sandboxes",
        "description": "List all sandboxes for the authenticated tenant.",
        "security": [
          {
            "TenantApiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Filter by sandbox status"
            },
            "required": false,
            "description": "Filter by sandbox status",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Maximum number of results",
              "example": "50"
            },
            "required": false,
            "description": "Maximum number of results",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Pagination cursor"
            },
            "required": false,
            "description": "Pagination cursor",
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of sandboxes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sandboxes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Sandbox"
                      }
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "total_count": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "sandboxes",
                    "next_cursor",
                    "total_count"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{sandboxId}": {
      "get": {
        "tags": [
          "Sandboxes"
        ],
        "summary": "Get sandbox details",
        "description": "Retrieve details for a specific sandbox.",
        "security": [
          {
            "TenantApiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Sandbox ID"
            },
            "required": true,
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Sandbox details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            }
          },
          "404": {
            "description": "Sandbox not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sandboxes"
        ],
        "summary": "Destroy a sandbox",
        "description": "Permanently destroy a sandbox and release its resources.",
        "security": [
          {
            "TenantApiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Sandbox ID"
            },
            "required": true,
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Sandbox destroyed"
          },
          "404": {
            "description": "Sandbox not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Sandbox state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sandboxes/{sandboxId}/files": {
      "post": {
        "tags": [
          "Sandboxes"
        ],
        "summary": "Write a file",
        "description": "Write a file to the sandbox filesystem.",
        "security": [
          {
            "TenantApiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Sandbox ID"
            },
            "required": true,
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteFileRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File written successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "size_bytes": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "path",
                    "size_bytes"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Sandbox not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{sandboxId}/files-list": {
      "get": {
        "tags": [
          "Sandboxes"
        ],
        "summary": "List files",
        "description": "List files in a directory within the sandbox.",
        "security": [
          {
            "TenantApiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Sandbox ID"
            },
            "required": true,
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Directory to list (default /workspace)"
            },
            "required": false,
            "description": "Directory to list (default /workspace)",
            "name": "directory",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "File listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Sandbox not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}